TradeJournalWPF/Converters/InverseBoolConverter.cs
using System;
using System.Globalization;
using System.Windows.Data;

namespace TradeJournalWPF.Converters
{
    public sealed class InverseBoolConverter : MarkupConverter, IValueConverter
    {
        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
        {
            if (targetType != typeof(bool))
                throw new InvalidOperationException("The target must be a boolean");

            return !(bool)value;
        }

        public new object ConvertBack(object value, Type targetType, object parameter,
            CultureInfo culture)
        {
            throw new NotSupportedException();
        }
    }
}
